Adding mvc model and authentication
