+++ .env
@@ -1,2 +1,3 @@
 DATABASE_URL="postgresql://Eyoba:Ihaveadream19@localhost:5432/gbi_gubae?schema=public"
 PORT=4500
+JWT_SECRET="This is the secret key"
