+++ middleware/adminAuth.js
@@ -1,0 +1,15 @@
+import jwt from "jsonwebtoken";
+
+function buildToken(res, studentid) {
+  const token = jwt.sign({ studentid }, process.env.JWT_SECRET, {
+    expiresIn: "3d",
+  });
+  res.cookie("jwt", token, {
+    httpObly: true,
+    secure: process.env.NODE_ENV !== "development",
+    sameSite: "strict",
+    maxAge: 30 * 24 * 60 * 60 * 1000,
+  });
+}
+
+export default buildToken;
