+++ server.js
@@ -1,36 +1,24 @@
 import express from "express";
 import cors from "cors";
-import add from "./routes/users/add-user.js";
-import getusers from "./routes/users/get-users.js";
-import getUser from "./routes/users/get-user.js";
-import updateUser from "./routes/users/update-user.js";
-import deleteUser from "./routes/users/delete-user.js";
 
-import addAdmin from "./routes/admin/add-admin.js";
-import deleteAdmin from "./routes/admin/delete-admin.js";
-import getAdmins from "./routes/admin/get-admins.js";
-import updateAdmin from "./routes/admin/update-admin.js";
-import getAdmin from "./routes/admin/get-admin.js";
+import UserRoutes from "./routes/user.routes.js";
+import AdminRoutes from "./routes/admin.routes.js";
 
-import login from "./routes/login.js";
+import { handleError } from "./utils/util.js";
+import cookieParser from "cookie-parser";
 
 const app = express();
 app.use(cors());
 app.use(express.json());
+app.use(cookieParser());
 
-app.use("/api/login", login);
+app.use("/api/user", UserRoutes);
+app.use("/api/admin", AdminRoutes);
 
-app.use("/api/create", add);
-app.use("/api/getAll", getusers);
-app.use("/api/get", getUser);
-app.use("/api/update", updateUser);
-app.use("/api/delete", deleteUser);
-
-app.use("/api/assign", addAdmin);
-app.use("/api/remove", deleteAdmin);
-app.use("/api/get-admins", getAdmins);
-app.use("/api/update-admin", updateAdmin);
-app.use("/api/get-admin", getAdmin);
+app.use((err, req, res, next) => {
+  const errorResponse = handleError(err);
+  res.status(errorResponse.statusCode || 500).json(errorResponse);
+});
 
 app.listen(process.env.PORT || 4500, () =>
   console.log("the server running on https://localhost:4500")
